+++ src/main/java/sonia/scm/jenkins/JenkinsRepositoryHookHandler.java
@@ -24,11 +24,8 @@
 
 package sonia.scm.jenkins;
 
-//~--- non-JDK imports --------------------------------------------------------
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import sonia.scm.net.ahc.AdvancedHttpClient;
 import sonia.scm.net.ahc.AdvancedHttpRequest;
 import sonia.scm.net.ahc.AdvancedHttpRequestWithBody;
@@ -36,91 +33,58 @@ import sonia.scm.net.ahc.AdvancedHttpResponse;
 import sonia.scm.net.ahc.BaseHttpRequest;
 import sonia.scm.repository.Changeset;
 import sonia.scm.repository.RepositoryHookEvent;
+import sonia.scm.util.HttpUtil;
+import sonia.scm.util.IOUtil;
 import sonia.scm.util.Util;
 
-//~--- JDK imports ------------------------------------------------------------
-
+import javax.inject.Provider;
 import java.io.IOException;
 import java.io.InputStream;
-
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-
 import java.util.List;
 import java.util.Set;
-import sonia.scm.util.IOUtil;
-
-import javax.inject.Provider;
 
 /**
- *
  * @author Sebastian Sdorra
  */
-public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
-{
+public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
 
-  /** Http parameter for jenkins authentication token */
   public static final String PARAMETER_TOKEN = "token";
+  private static final Logger logger = LoggerFactory.getLogger(JenkinsRepositoryHookHandler.class);
 
-  /** the logger for JenkinsRepositoryHookHandler */
-  private static final Logger logger =
-    LoggerFactory.getLogger(JenkinsRepositoryHookHandler.class);
+  private final JenkinsConfiguration configuration;
+  private final Provider<AdvancedHttpClient> httpClientProvider;
 
-  //~--- constructors ---------------------------------------------------------
-
-  /**
-   * Constructs ...
-   *
-   *
-   * @param httpClientProvider
-   * @param configuration
-   */
   public JenkinsRepositoryHookHandler(Provider<AdvancedHttpClient> httpClientProvider,
-                                      JenkinsConfiguration configuration)
-  {
+                                      JenkinsConfiguration configuration) {
     this.httpClientProvider = httpClientProvider;
     this.configuration = configuration;
   }
 
-  //~--- methods --------------------------------------------------------------
-
-  /**
-   * Method description
-   *
-   *
-   * @param event
-   */
   @Override
-  public void sendRequest(RepositoryHookEvent event)
-  {
-    if (configuration.getBranches().isEmpty())
-    {
+  public void sendRequest(RepositoryHookEvent event) {
+    if (configuration.getBranches().isEmpty()) {
       logger.debug("branch list is empty, send request");
       handleRepositoryEvent(configuration);
-    }
-    else
-    {
-      if (isInBranchSet(configuration.getBranches(), event.getContext().getChangesetProvider().getChangesets()))
-      {
+    } else {
+      if (isInBranchSet(configuration.getBranches(), event.getContext().getChangesetProvider().getChangesets())) {
         handleRepositoryEvent(configuration);
-      }
-      else
-      {
+      } else {
         logger.debug("changesets does not contain configured branches");
       }
     }
   }
-  
-  private String createBaseUrl(JenkinsConfiguration configuration){
+
+  private String createBaseUrl(JenkinsConfiguration configuration) {
     String url = configuration.getUrl();
 
-    if (!url.endsWith("/"))
-    {
+    if (!url.endsWith("/")) {
       url = url.concat("/");
     }
-    
+
     return url;
   }
 
@@ -129,19 +93,17 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
    * visible to the package for testing.
    *
    * @param configuration jenkins configuration
-   *
    * @return the url to the remote trigger servlet of jenkins
    */
-  String createUrl(JenkinsConfiguration configuration)
-  {
+  String createUrl(JenkinsConfiguration configuration) {
     String url = createBaseUrl(configuration);
 
     //J-
     // url encode proejct name, see http://goo.gl/v8Rond
     return escape(
       url.concat("job/")
-         .concat(configuration.getProject())
-         .concat("/build")
+        .concat(configuration.getProject())
+        .concat("/build")
     );
     //J+
   }
@@ -152,33 +114,25 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
    * @param urlString url
    * @return escaped url
    */
-  String escape(String urlString)
-  {
+  String escape(String urlString) {
     String escapedUrl = urlString;
 
-    try
-    {
+    try {
       URL url = new URL(urlString);
       //J-
       URI uri = new URI(
-        url.getProtocol(), 
-        url.getUserInfo(), 
-        url.getHost(), 
-        url.getPort(), 
-        url.getPath(), 
-        url.getQuery(), 
+        url.getProtocol(),
+        url.getUserInfo(),
+        url.getHost(),
+        url.getPort(),
+        url.getPath(),
+        url.getQuery(),
         url.getRef()
       );
       //J+
 
       escapedUrl = uri.toString();
-    }
-    catch (URISyntaxException ex)
-    {
-      logger.warn("could not escaped url", ex);
-    }
-    catch (MalformedURLException ex)
-    {
+    } catch (URISyntaxException | MalformedURLException ex) {
       logger.warn("could not escaped url", ex);
     }
 
@@ -191,115 +145,86 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
    *
    * @param configuration jenkins configuration
    */
-  private void handleRepositoryEvent(JenkinsConfiguration configuration)
-  {
+  private void handleRepositoryEvent(JenkinsConfiguration configuration) {
 
-    // create the url to the endpoint
     String url = createUrl(configuration);
 
-    // log the url
-    if (logger.isInfoEnabled())
-    {
+    if (logger.isInfoEnabled()) {
       logger.info("call jenkins at {}", url);
     }
 
-    try
-    {
-
-      // send the request to the jenkins server
+    try {
       sendRequest(configuration, url);
-    }
-
-    // catch each IOException and log them
-    catch (IOException ex)
-    {
+    } catch (IOException ex) {
       logger.error("could not send request to jenkins", ex);
     }
   }
-  
-  private String createCrumbUrl(JenkinsConfiguration configuration)
-  {
+
+  private String createCrumbUrl(JenkinsConfiguration configuration) {
     return escape(createBaseUrl(configuration).concat("crumbIssuer/api/xml"));
   }
-  
-  private CsrfCrumb parseCsrfCrumbResponse(AdvancedHttpResponse response) throws IOException
-  {
-    CsrfCrumb csrfCrumb = null;
+
+  private CsrfCrumb parseCsrfCrumbResponse(AdvancedHttpResponse response) throws IOException {
+    CsrfCrumb csrfCrumb;
     InputStream content = null;
-    try 
-    {
+    try {
       content = response.contentAsStream();
       csrfCrumb = CsrfCrumbParser.parse(content);
-    } 
-    finally 
-    {
+    } finally {
       IOUtil.close(content);
     }
     return csrfCrumb;
   }
-  
-  private CsrfCrumb getJenkinsCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client)
-  {
+
+  private CsrfCrumb getJenkinsCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client) {
     String url = createCrumbUrl(configuration);
     logger.debug("fetch csrf crumb from {}", url);
 
     AdvancedHttpRequest request = client.get(url);
 
     appendAuthenticationHeader(configuration, request);
-    
+
     CsrfCrumb crumb = null;
-    try
-    {
+    try {
       AdvancedHttpResponse response = request.request();
       int sc = response.getStatus();
-      if (sc != 200)
-      {
+      if (sc != 200) {
         logger.warn("jenkins crumb endpoint returned status code {}", sc);
-      } 
-      else 
-      {
+      } else {
         crumb = parseCsrfCrumbResponse(response);
       }
-    }
-    catch (IOException ex)
-    {
+    } catch (IOException ex) {
       logger.warn("failed to fetch csrf crumb", ex);
     }
-    
+
     return crumb;
   }
-  
-  private void appendCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client, BaseHttpRequest request)
-  {
-    if (configuration.isCsrf())
-    {
+
+  private void appendCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client, BaseHttpRequest request) {
+    if (configuration.isCsrf()) {
       CsrfCrumb crumb = getJenkinsCsrfCrumb(configuration, client);
-      logger.debug("add csrf crumb to api request");
-      request.header(crumb.getCrumbRequestField(), crumb.getCrumb());
-    }
-    else 
-    {
+      if (crumb != null) {
+        logger.debug("add csrf crumb to api request");
+        request.header(crumb.getCrumbRequestField(), crumb.getCrumb());
+      }
+    } else {
       logger.debug("csrf protection is disabled, skipping crumb");
     }
   }
-  
-  private void appendAuthenticationHeader(JenkinsConfiguration configuration, BaseHttpRequest request){
+
+  private void appendAuthenticationHeader(JenkinsConfiguration configuration, BaseHttpRequest request) {
     // check for authentication parameters
     String username = configuration.getUsername();
     String apiToken = configuration.getApiToken();
 
-    if (Util.isNotEmpty(username) && Util.isNotEmpty(apiToken))
-    {
-      if (logger.isDebugEnabled())
-      {
+    if (Util.isNotEmpty(username) && Util.isNotEmpty(apiToken)) {
+      if (logger.isDebugEnabled()) {
         logger.debug("added authentication for user {}", username);
       }
 
       // add basic authentication header
       request.basicAuth(username, apiToken);
-    }
-    else if (logger.isDebugEnabled())
-    {
+    } else if (logger.isDebugEnabled()) {
       logger.debug("skip authentication. username or api token is empty");
     }
   }
@@ -308,31 +233,27 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
    * Send the request to the jenkins ci server to trigger a new build.
    *
    * @param configuration jenkins configuration
-   * @param url url of the jenkins server
-   *
+   * @param url           url of the jenkins server
    * @throws IOException
    */
   private void sendRequest(JenkinsConfiguration configuration, String url)
-    throws IOException
-  {
+    throws IOException {
 
     /**
      * Create a new http client from the Guice Provider.
      */
     AdvancedHttpClient httpClient = httpClientProvider.get();
+    url = appendBuildParameters(configuration, url);
 
     // retrive authentication token
     AdvancedHttpRequestWithBody request = httpClient.post(url);
     String token = configuration.getToken();
     // check if the token is not empty.
-    if (Util.isNotEmpty(token))
-    {
+    if (Util.isNotEmpty(token)) {
 
       // add the token as parameter for the request
       request.queryString(PARAMETER_TOKEN, token);
-    }
-    else if (logger.isDebugEnabled())
-    {
+    } else if (logger.isDebugEnabled()) {
       logger.debug("no project token is available");
     }
 
@@ -345,83 +266,54 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler
      */
     int sc = request.request().getStatus();
 
-    // if the response is greater than 400 write an error to the log
-    if (sc >= 400)
-    {
+    if (sc >= 400) {
       logger.error("jenkins returned status code {}", sc);
-    }
-    else if (logger.isInfoEnabled())
-    {
-      logger.info("jenkins hook successfully submited");
+    } else if (logger.isInfoEnabled()) {
+      logger.info("jenkins hook successfully submitted");
     }
   }
 
-  //~--- get methods ----------------------------------------------------------
+  private String appendBuildParameters(JenkinsConfiguration configuration, String url) {
+    Set<BuildParameter> buildParameters = configuration.getBuildParameters();
 
-  /**
-   * Method description
-   *
-   *
-   * @param branchSet
-   * @param changesets
-   *
-   * @return
-   */
+    if (!buildParameters.isEmpty()) {
+      StringBuilder builder = new StringBuilder(url);
+      builder.append("WithParameters?");
+      buildParameters.forEach(
+        parameter -> builder
+          .append(HttpUtil.encode(parameter.getName()))
+          .append("=")
+          .append(HttpUtil.encode(parameter.getValue()))
+          .append("&")
+      );
+      url = builder.toString().substring(0, builder.toString().length() - 1);
+    }
+
+    return url;
+  }
+
   private boolean isInBranchSet(Set<String> branchSet,
-    Iterable<Changeset> changesets)
-  {
-    boolean found = false;
-
-    for (Changeset changeset : changesets)
-    {
-      if (isInBranchSet(branchSet, changeset))
-      {
-        found = true;
-
-        break;
+                                Iterable<Changeset> changesets) {
+    for (Changeset changeset : changesets) {
+      if (isInBranchSet(branchSet, changeset)) {
+        return true;
       }
     }
-
-    return found;
+    return false;
   }
 
-  /**
-   * Method description
-   *
-   *
-   * @param branchSet
-   * @param changeset
-   *
-   * @return
-   */
-  private boolean isInBranchSet(Set<String> branchSet, Changeset changeset)
-  {
-    boolean found = false;
+  private boolean isInBranchSet(Set<String> branchSet, Changeset changeset) {
     List<String> branches = changeset.getBranches();
 
-    if (Util.isNotEmpty(branches))
-    {
-      for (String branch : branches)
-      {
-        if (branchSet.contains(branch))
-        {
-          logger.debug("found branch {} at {}, send request", branch,
-            changeset.getId());
-          found = true;
-
-          break;
+    if (Util.isNotEmpty(branches)) {
+      for (String branch : branches) {
+        if (branchSet.contains(branch)) {
+          logger.debug("found branch {} at {}, send request", branch, changeset.getId());
+          return true;
         }
       }
     }
 
-    return found;
+    return false;
   }
-
-  //~--- fields ---------------------------------------------------------------
-
-  /** Field description */
-  private JenkinsConfiguration configuration;
-
-  /** Guice provider for {@link AdvancedHttpClient} */
-  private Provider<AdvancedHttpClient> httpClientProvider;
 }